replaced align_csv with csvaddcol script
The new script can be used in a unix pipeline with csvkit and other
unix utilities (sed etc) to align a given csv to a deriva catalog.
The new script adds just the missing functionality to that pipeline
of adding columns to a csv based on results of queries to ermrest.